ui/backend/server/requirements.txt
aioboto3==11.3.0
aiobotocore==2.6.0
aiofiles==23.2.1
aiohttp==3.8.6
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
boto3==1.28.17
botocore==1.31.17
bytecode==0.15.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorlog==6.7.0
cryptography==41.0.4
ddsketch==2.0.4
ddtrace==1.15.0
Deprecated==1.2.14
Django==4.2.6
django-extensions==3.2.3
django-ninja==1.0b1
envier==0.4.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
jmespath==1.0.1
multidict==6.0.4
opentelemetry-api==1.20.0
passlib==1.7.4
propelauth-py==3.1.8
protobuf==4.24.4
psutil==5.9.5
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.17
wrapt==1.15.0
xmltodict==0.13.0
yarl==1.9.2
zipp==3.17.0
sf-hamilton==1.58.0
